// src/app/models/Lancamento.ts
export interface Lancamento{
    recibo: string
    data_lan: Date
    data_ven: Date
    tipo_doc: string
    num_doc: string
    entrada: string
    saida: string
    cong: string
    forn: string
    dizimista: string
    obs: string
    valor: string
    conta: string
    situacao: string
    tipo_lanc: string
    historico: string
    status_lanc: string
}